+++ attendance_dump_20250808_115319.json
@@ -1,0 +1,156 @@
+[
+  {
+    "id": 1,
+    "emp": 2,
+    "emp_code": "1",
+    "first_name": "Nidhin",
+    "last_name": null,
+    "department": "IT",
+    "position": "consultant",
+    "punch_time": "2025-08-04 19:14:21",
+    "punch_state": "1",
+    "punch_state_display": "Check Out",
+    "verify_type": 10,
+    "verify_type_display": "Fingerprint & Card",
+    "work_code": "0",
+    "gps_location": null,
+    "area_alias": "AREA1",
+    "terminal_sn": "QEM1243700149",
+    "temperature": 0.0,
+    "is_mask": "No",
+    "terminal_alias": "zkta",
+    "upload_time": "2025-08-04 14:14:22"
+  },
+  {
+    "id": 2,
+    "emp": 2,
+    "emp_code": "1",
+    "first_name": "Nidhin",
+    "last_name": null,
+    "department": "IT",
+    "position": "consultant",
+    "punch_time": "2025-08-04 19:06:44",
+    "punch_state": "0",
+    "punch_state_display": "Check In",
+    "verify_type": 1,
+    "verify_type_display": "Fingerprint",
+    "work_code": "0",
+    "gps_location": null,
+    "area_alias": "AREA1",
+    "terminal_sn": "QEM1243700149",
+    "temperature": 0.0,
+    "is_mask": "No",
+    "terminal_alias": "zkta",
+    "upload_time": "2025-08-04 20:12:13"
+  },
+  {
+    "id": 3,
+    "emp": 2,
+    "emp_code": "1",
+    "first_name": "Nidhin",
+    "last_name": null,
+    "department": "IT",
+    "position": "consultant",
+    "punch_time": "2025-08-04 19:08:09",
+    "punch_state": "0",
+    "punch_state_display": "Check In",
+    "verify_type": 1,
+    "verify_type_display": "Fingerprint",
+    "work_code": "0",
+    "gps_location": null,
+    "area_alias": "AREA1",
+    "terminal_sn": "QEM1243700149",
+    "temperature": 0.0,
+    "is_mask": "No",
+    "terminal_alias": "zkta",
+    "upload_time": "2025-08-04 20:12:13"
+  },
+  {
+    "id": 4,
+    "emp": 2,
+    "emp_code": "1",
+    "first_name": "Nidhin",
+    "last_name": null,
+    "department": "IT",
+    "position": "consultant",
+    "punch_time": "2025-08-04 19:09:11",
+    "punch_state": "0",
+    "punch_state_display": "Check In",
+    "verify_type": 10,
+    "verify_type_display": "Fingerprint & Card",
+    "work_code": "0",
+    "gps_location": null,
+    "area_alias": "AREA1",
+    "terminal_sn": "QEM1243700149",
+    "temperature": 0.0,
+    "is_mask": "No",
+    "terminal_alias": "zkta",
+    "upload_time": "2025-08-04 20:12:13"
+  },
+  {
+    "id": 5,
+    "emp": 2,
+    "emp_code": "1",
+    "first_name": "Nidhin",
+    "last_name": null,
+    "department": "IT",
+    "position": "consultant",
+    "punch_time": "2025-08-04 19:52:35",
+    "punch_state": "0",
+    "punch_state_display": "Check In",
+    "verify_type": 10,
+    "verify_type_display": "Fingerprint & Card",
+    "work_code": "0",
+    "gps_location": null,
+    "area_alias": "AREA1",
+    "terminal_sn": "QEM1243700149",
+    "temperature": 0.0,
+    "is_mask": "No",
+    "terminal_alias": "zkta",
+    "upload_time": "2025-08-04 20:12:13"
+  },
+  {
+    "id": 6,
+    "emp": 3,
+    "emp_code": "2",
+    "first_name": "Nid",
+    "last_name": "V",
+    "department": "IT",
+    "position": "consultant",
+    "punch_time": "2025-08-04 20:14:50",
+    "punch_state": "0",
+    "punch_state_display": "Check In",
+    "verify_type": 1,
+    "verify_type_display": "Fingerprint",
+    "work_code": "0",
+    "gps_location": null,
+    "area_alias": "AREA1",
+    "terminal_sn": "QEM1243700149",
+    "temperature": 0.0,
+    "is_mask": "No",
+    "terminal_alias": "zkta",
+    "upload_time": "2025-08-04 20:14:51"
+  },
+  {
+    "id": 7,
+    "emp": 3,
+    "emp_code": "2",
+    "first_name": "Nid",
+    "last_name": "V",
+    "department": "IT",
+    "position": "consultant",
+    "punch_time": "2025-08-04 20:16:46",
+    "punch_state": "0",
+    "punch_state_display": "Check In",
+    "verify_type": 10,
+    "verify_type_display": "Fingerprint & Card",
+    "work_code": "0",
+    "gps_location": null,
+    "area_alias": "AREA1",
+    "terminal_sn": "QEM1243700149",
+    "temperature": 0.0,
+    "is_mask": "No",
+    "terminal_alias": "zkta",
+    "upload_time": "2025-08-04 20:16:47"
+  }
+]
