+++ logs1/status.json
@@ -1,0 +1,1 @@
+{"lift_off_timestamp":"2025-08-04 20:19:05.805006","mission_accomplished_timestamp":"2025-08-04 20:19:05.874792"}
