+++ logs2/status.json
@@ -1,0 +1,1 @@
+{"lift_off_timestamp":"2025-08-08 10:28:39.068000","mission_accomplished_timestamp":"2025-08-08 10:28:39.438993"}
